Moves hook to hook folder
